Added box describing each file
- each file has a box at the top of the file describing:
    1. the purpose of the file
    2. what is contained in the file
