+++ assets/assets.js
@@ -1,3 +1,17 @@
+/////////////////////////////////////////////////////////
+//                                                     //
+// Purpose: Defines constants within the app           //
+//                                                     //
+// What it contains:                                   //
+//      - Colours                                      //
+//      - Strings                                      //
+//      - Enumerations                                 //
+//                                                     //
+/////////////////////////////////////////////////////////
+
+
+
+
 // =================================================
 // == Colours                                     ==
 // =================================================
