+++ index.html
@@ -1,3 +1,18 @@
+<!------------------------------------------------
+
+Purpose: Defines static visuals for the program
+
+What it contains:
+    - description text for the program
+    - the buttons for the graphs
+    - dropdown selections for the graphs
+    - empty tables and empty SVGs for the graphs
+    - different external dependencies
+    - link to the overall main function of the program
+
+-------------------------------------------------->
+
+
 <!DOCTYPE html><!--[if lt IE 9]>
     <html class="no-js lt-ie9" lang="en" dir="ltr"><![endif]--><!--[if gt IE 8]><!-->
     <html class="no-js" lang="en" dir="ltr">
