--- model/dataModels/dataModels.js
@@ -1,90 +1,0 @@
-import { AgeSexGroupOrder } from "../../assets/assets.js";
-
-
-const SortedAgeSexGroupHeadings = Object.keys(AgeSexGroupOrder).sort((a,b) => {return AgeSexGroupOrder[a] - AgeSexGroupOrder[b]});
-
-
-// CSVDataModel: Data model for importing data from CSV files
-export class CSVDataModel {
-    constructor(data) {
-        this.data = data;
-    }
-}
-
-
-// FoodIngredientDataModel: Data model for the food ingredient data in the CSV file
-export class FoodIngredientDataModel extends CSVDataModel {
-    constructor(data, dataGroupedByNutrientAndDemoList, dataGroupedByNutrientAndDemo, fullyNestedDataByFoodGroup) {
-        super(data);
-        this.dataGroupedByNutrientAndDemoList = dataGroupedByNutrientAndDemoList;
-        this.dataGroupedByNutrientAndDemo = dataGroupedByNutrientAndDemo;
-        this.fullyNestedDataByFoodGroup = fullyNestedDataByFoodGroup;
-
-        this.ageSexGroupHeadings = SortedAgeSexGroupHeadings;
-    }
-
-    getNutrientData(nutrient) {
-        return this.dataGroupedByNutrientAndDemo[nutrient];
-    }
-
-    // calculate the total amount by nutrient per age-sex group
-    findNutrientTotalAmtPerAgeSexGroup(nutrientData, graphType) {
-        let maxAccumulatedAmount = 0;
-
-        /* If graph type is number, get data from "Amount" col, otherwise use "Percentage" */
-        const keyName = graphType === "number" ? "Amount" : "Percentage"
-
-        const groupedAmount = Object.keys(nutrientData).reduce((obj, ageSexGroup) => {
-            obj[ageSexGroup] = Object.keys(nutrientData[ageSexGroup]).reduce((innerObj, foodLevelGroup) => {
-                innerObj[foodLevelGroup] = nutrientData[ageSexGroup][foodLevelGroup].reduce((amount, dataRow) => {
-                    // we want only the overall amounts for each food group, so ignore the rows representing subgroups
-                    return amount + (isNaN(dataRow[keyName]) || dataRow["Food group_level2"] || dataRow["Food group_level3"] ? 0 : dataRow[keyName])
-                }, 0);
-                return innerObj;
-            }, {})
-            maxAccumulatedAmount = Math.max(maxAccumulatedAmount, Object.values(obj[ageSexGroup]).reduce((sum, cur) => sum + cur, 0))
-            return obj;
-        }, {})
-
-        return {"groupedAmount": groupedAmount, "maxAccumulatedAmount": maxAccumulatedAmount};
-    }
-
-    buildSunBurstTree(nutrient, ageSexGroup) {
-        const nutrientData = this.fullyNestedDataByFoodGroup[nutrient];
-
-        /* Group data into a tree where each node has the form of { name, value, row, children } for d3.hierarchy() */
-        let groupedPercentages = Object.keys(nutrientData[ageSexGroup]).reduce((objLevel1, foodLevel1) => {
-            const foodLevel1Group = nutrientData[ageSexGroup][foodLevel1];
-
-            objLevel1.children.push(Object.keys(nutrientData[ageSexGroup][foodLevel1]).filter(d => d).reduce((objLevel2, foodLevel2) => {
-                const foodLevel2Group = foodLevel1Group[foodLevel2];
-                objLevel2.value -= foodLevel2Group[""]["Amount"];
-                const newChild = {
-                    name: foodLevel2,
-                    value: foodLevel2Group[""]["Amount"],
-                    row: foodLevel2Group[""]
-                };
-                newChild.children = Object.keys(foodLevel2Group).filter(d => d).map((foodLevel3) => {
-                    const foodLevel3Group = foodLevel2Group[foodLevel3];
-                    newChild.value -= foodLevel3Group["Amount"];
-                    return {
-                        name: foodLevel3,
-                        value: foodLevel3Group["Amount"],
-                        row: foodLevel3Group
-                    }
-                })
-                objLevel2.children.push(newChild);
-                return objLevel2;
-            }, { 
-                name: foodLevel1, 
-                value: foodLevel1Group[""][""]["Amount"], // key "" represents the overall group including all subgroups
-                row: foodLevel1Group[""][""], 
-                children: []
-            }));
-            return objLevel1;
-        }, { name: "All Items", row: {Percentage: 100}, children: [] });
-
-        groupedPercentages = {name: "data", children: [groupedPercentages]}
-        return groupedPercentages;
-    }
-}
