+++ viewController/viewController.js
@@ -1,8 +1,21 @@
+////////////////////////////////////////////////////////////////////
+//                                                                //
+// Purpose: Handles the overall display of the program            //
+//                                                                //
+// What it Contains:                                              //
+//      - wrappers to draw and update the sun burst graph and     //
+//          the bar graph                                         //
+//      - the main function to run the overall program            //
+//      - initiliazes the backend of the program                  //
+//                                                                //
+////////////////////////////////////////////////////////////////////      
+
+
+
 import { Model } from '../model/model.js'
 import { BarGraph } from './components/barGraph.js';
 import { SunBurst } from './components/sunBurstGraph.js';
-import { ViewTools } from './tools/viewTools.js';
-import { TranslationTools, TranslationObj } from '../tools/translationTools.js';
+import { TranslationTools, TranslationObj, ViewTools } from '../tools/tools.js';
 import { Component } from './components/component.js';
